refactor: split large service files into domain modules (#336)
