+++ apps/node-agent/src/services/agentService/agent-commands.ts
@@ -1,0 +1,1 @@
+export { AgentService } from './agent-lifecycle';
